+++ src/main/kotlin/dev/micropythontools/ui/MpyFileSystemWidget.kt
@@ -494,7 +494,10 @@ internal class FileSystemWidget(private val project: Project) : JBPanel<FileSyst
             )
 
         try {
-            dirList = deviceService.blindExecute(fileSystemScanScript)
+            // If not using the reporter, this refresh is a part of some complex operation that needs
+            // information about the file system state, such as uploads
+            // In that scenario it shouldn't return to the connected state to prevent emitting extra MicroPython banners
+            dirList = deviceService.blindExecute(fileSystemScanScript, !useReporter)
         } catch (e: CancellationException) {
             if (disconnectOnCancel) {
                 deviceService.disconnect(reporter)